requirements.txt
PyYAML
jupyter
ipython
chardet
